AI知识库+企微客服相关代码提交

+++ src/main/java/cn/iyque/service/IYqueEmbeddingService.java
@@ -35,5 +35,5 @@ public interface IYqueEmbeddingService {
      * @param kid
      * @return
      */
-    List<Double> getQueryVector(String query, String kid);
+    List<Float> getQueryVector(String query, String kid);
 }
